\name{getAssemblies}
\alias{getAssemblies}

\title{
 Extracting Assemblies.
}

\description{
  This function extracts the assemblies from an xml object.
}
\usage{
  getAssemblies(xml)
}

\arguments{
  \item{xml}{An xml object.}
}

\details{
 This function extracts the information from an imported xml object.
}
\value{
  A matrix.
}

\author{
  Daniel Fischer
}

\examples{
\dontrun{
getAssemblies(xml)
}
}

\keyword{methods}
 
