\name{getGeneLocation}
\alias{getGeneLocation}

\title{
 Extracting Gene Locations.
}

\description{
  This function extracts the gene locations from an imported gtf file.
}
\usage{
  getGeneLocation(gtf)
}

\arguments{
  \item{gtf}{An imported gtf object.}
}

\details{
 This function extracts the information from an imported gtf object.
}
\value{
  A matrix.
}

\author{
  Daniel Fischer
}

\examples{
\dontrun{
getGeneLocation(gtf)
}
}

\keyword{methods}
 
