\name{getSequenceFromNCBI}
\alias{getSequenceFromNCBI}

\title{
Extracts a sequence from the NCBI webpage
}

\description{
Retrieve a sequence from the NCBI webpage
}
\usage{
  getSequenceFromNCBI(id, file=NULL)
}

\arguments{
  \item{id}{The gene identifier}
  \item{file}{File name to where the sequence shall be stored}
}

\details{
This function extracts the sequence for a given identifer and then stores, if requested the sequence to the HDD.
 
}

\author{
  Daniel Fischer
}

\keyword{methods}
 
