\name{hoardeR-package}
\alias{R/hoardeR-package}
\docType{package}

\title{
  Collect and Retrieve Annotation Data for Various Genetic Data.
}

\description{
  The hoardeR package is designed for collecting and retrieving data from
  various sources. The current main focus is on setting up a connection to
  the NCBI Blast service. Also, the gene information for Ensemble Genes can
  be retrieved from NCBI.
}

\details{
  \tabular{ll}{
    Package: \tab hoardeR\cr
    Type: \tab Package\cr
    Version: \tab 0.0-1\cr
    Date: \tab 2014-02-20\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Daniel Fischer

  Maintainer: Daniel Fischer <daniel.fischer@mtt.fi>
}

\keyword{multivariate}

