\name{importBlastTab}
\alias{importBlastTab}

\title{
  Import a Tab Delimited Blast Output File
}

\description{
  This function imports a tab delimited blast output.
}
\usage{
  importBlastTab(file)
}

\arguments{
  \item{file}{File name of the file.}
}

\details{
  This function imports a tab delimited blast output file, currently the same as \code{read.table}
}
\value{
  A data frame containing the columns of the file.
}

\author{
  Daniel Fischer
}

 
