\name{plotCoverage}
\alias{plotCoverage}

\title{
Plots a coverage density object
}

\description{
Plots a coverage density object.
}
\usage{
  plotCoverage(x, use.sqrt=TRUE)
}

\arguments{
  \item{x}{A coverage density object}
  \item{use.sqrt}{Logical, use sqrt scale?}
}

\details{
 This function plots the coverage of bam-files
 
}

\author{
  Daniel Fischer
}

\keyword{methods}
 
