\name{print.xmlImport}
\docType{methods}
\alias{print,xmlImport-method}
\alias{print.xmlImport}

\title{Print an xmlImport Object}
\description{
  Prints an \code{xmlImport} object.
}
\usage{
 \method{print}{xmlImport}(x, n=2, ...)
}

\arguments{
  \item{x}{Object of class \code{xmlImport}.}
  \item{n}{Amount of elements to be displayed, numeric.}
  \item{...}{Additional parameters.}
}

\details{
The print function displays an \code{xmlImport} object. By default just the first two elements are displayed.
}

\author{Daniel Fischer}

\keyword{methods}
\keyword{print}
