\name{species}
\alias{species}
\title{
Species Name for Blast Search
}
\description{
Standardized species names for blast search.
}
\usage{
species
}
\format{
  This vector contains 134 species names.
}
\note{
 Note, the names have been extracted on 1.6.2014 from the NCBI server.
}
\keyword{datasets}
 
