\name{summary.ensgInfo}
\docType{methods}
\alias{summary.ensgInfo}
\alias{summary,ensgInfo-method}

\title{Summarize an ensgInfo Object}
\description{
  Summarizes and prints an \code{ensgInfo} object in an informative way.
}
\usage{
 \method{summary}{ensgInfo}(object, ...)
}

\arguments{
  \item{object}{Object of class \code{ensgInfo}.}
  \item{...}{Additional parameters.}
}

\details{
 Summary for a \code{ensgInfo} object, providing the amount of different gene types in the query.
 }

\author{Daniel Fischer}

\keyword{methods}
\keyword{print}
