\name{summary.fa}
\docType{methods}
\alias{summary.fa}
\alias{summary,fa-method}

\title{Summarize an fa Object}
\description{
  Summarizes and prints an \code{fa} object in an informative way.
}
\usage{
 \method{summary}{fa}(object, ...)
}

\arguments{
  \item{object}{Object of class \code{fa}.}
  \item{...}{Additional parameters.}
}

\details{
 Summary for a \code{fa} object, providing the amount of sequences, the minimum and maximum length as well as the
 average length.
}

\author{Daniel Fischer}

\keyword{methods}
\keyword{print}
