\name{hoardeR-package}
\alias{R/hoardeR-package}
\docType{package}

\title{
  Collect and Retrieve Annotation Data for Various Genetic Data.
}

\description{
  The hoardeR package is designed for collecting, retrieving and transforming data
  from various sources. The current main focus is on setting up a connection to
  the NCBI Blast service. Also, the gene information for Ensemble Genes can
  be retrieved from NCBI.
  Methods for visualizing the results are currently under development. The latest
  'night-build' of the package can be retrieved from
  
  https://github.com/fischuu/hoardeR
}

\details{
  \tabular{ll}{
    Package: \tab hoardeR\cr
    Type: \tab Package\cr
    Version: \tab 0.2\cr
    Date: \tab 2014-10-06\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Daniel Fischer, Anu Sironen

  Maintainer: Daniel Fischer <daniel.fischer@mtt.fi>
}

\keyword{multivariate}

