\name{hoardeR-package}
\alias{R/hoardeR-package}
\docType{package}

\title{
  Collect and Retrieve Annotation Data for Various Genomic Data Using Different Webservices.
}

\description{
  The hoardeR package is designed for collecting, retrieving and transforming data
  from various sources. The current main focus is on setting up a connection to
  the NCBI Blast service. Also, the gene information for Ensembl Genes can
  be retrieved from NCBI.
  Methods for visualizing the results are are also provided. The latest
  developer version of the package can be downloaded from
  
  https://github.com/fischuu/hoardeR
}

\details{
  \tabular{ll}{
    Package: \tab hoardeR\cr
    Type: \tab Package\cr
    Version: \tab 0.9\cr
    Date: \tab 2016-08-29\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Daniel Fischer, Anu Sironen

  Maintainer: Daniel Fischer <daniel.fischer@luke.fi>
}

\keyword{multivariate}

